+++ ethers-abigen/src/main/kotlin/io/ethers/abigen/AbiTypeParameter.kt
@@ -127,16 +127,17 @@ sealed interface AbiTypeParameter {
                     PropertySpec.builder(
                         StructFactory<*>::abi.name,
                         AbiType.Struct::class.asClassName().parameterizedBy(className),
-                    )
-                        .addModifiers(KModifier.OVERRIDE)
-                        .addAnnotation(JvmStatic::class)
-                        .initializer(
-                            "%T(${className.simpleName}::class, ::fromTuple, %L)",
-                            AbiType.Struct::class,
-                            getFieldAbiInitializers(),
-                        )
-                        .build(),
                 )
+                    .addModifiers(KModifier.OVERRIDE)
+                    .addAnnotation(JvmStatic::class)
+                    .initializer(
+                        "%T(%T::class.java, this, %L)",
+                        AbiType.Struct::class,
+                        className,
+                        getFieldAbiInitializers(),
+                    )
+                    .build(),
+            )
                 .addFunction(
                     FunSpec.builder(StructFactory<*>::fromTuple.name)
                         .addAnnotation(JvmStatic::class)