+++ ethers-crypto/src/test/kotlin/io/ethers/crypto/bip39/MnemonicCodeTest.kt
@@ -1,7 +1,8 @@
 package io.ethers.crypto.bip39
 
-import com.fasterxml.jackson.databind.json.JsonMapper
-import com.fasterxml.jackson.module.kotlin.KotlinModule
+import com.fasterxml.jackson.annotation.JsonCreator
+import com.fasterxml.jackson.annotation.JsonProperty
+import io.ethers.core.Jackson
 import io.kotest.assertions.throwables.shouldThrow
 import io.kotest.core.spec.style.FunSpec
 import io.kotest.inspectors.forAll
@@ -46,10 +47,8 @@ class MnemonicCodeTest : FunSpec({
     }
 
     context("test vectors") {
-        val mapper = JsonMapper.builder().addModule(KotlinModule.Builder().build()).build()
-
         test("english") {
-            val vectors = mapper.readerForListOf(TestVector::class.java).readValue<List<TestVector>>(
+            val vectors = Jackson.MAPPER.readerForListOf(TestVector::class.java).readValue<List<TestVector>>(
                 javaClass.getResource("/bip39/wordlist_english_test_vector.json"),
             )
 
@@ -63,11 +62,11 @@ class MnemonicCodeTest : FunSpec({
         }
     }
 }) {
-    private data class TestVector(
-        val entropy: String,
-        val mnemonic: String,
-        val passphrase: String,
-        val seed: String,
-        val bip32_xprv: String,
+    private data class TestVector @JsonCreator constructor(
+        @param:JsonProperty("entropy") val entropy: String,
+        @param:JsonProperty("mnemonic") val mnemonic: String,
+        @param:JsonProperty("passphrase") val passphrase: String,
+        @param:JsonProperty("seed") val seed: String,
+        @param:JsonProperty("bip32_xprv") val bip32Xprv: String,
     )
 }