+++ ethers-ens/src/main/kotlin/io/ethers/ens/Avatar.kt
@@ -1,6 +1,8 @@
 package io.ethers.ens
 
+import com.fasterxml.jackson.annotation.JsonCreator
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties
+import com.fasterxml.jackson.annotation.JsonProperty
 import io.ethers.core.Result
 import io.ethers.core.failure
 import io.ethers.core.success
@@ -10,7 +12,10 @@ import java.math.BigInteger
 import java.net.URI
 
 @JsonIgnoreProperties(ignoreUnknown = true)
-internal data class MetadataDTO(val image: String)
+internal data class MetadataDTO @JsonCreator constructor(
+    @param:JsonProperty("image")
+    val image: String,
+)
 
 internal enum class AvatarNFTType {
     ERC721,